+++ src/System.Diagnostics.StackTrace/src/System/Diagnostics/StackTraceSymbols.cs
