+++ src/System.Reflection.TypeExtensions/tests/CoreCLR/ModuleTests.cs
