+++ src/System.Reflection.TypeExtensions/tests/CoreCLR/System.Reflection.TypeExtensions.CoreCLR.Tests.csproj
@@ -10,7 +10,7 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MemberInfoTests.CoreCLR.cs" />
-    <Compile Include="ModuleTests.CoreCLR.cs" />
+    <Compile Include="MemberInfoTests.cs" />
+    <Compile Include="ModuleTests.cs" />
   </ItemGroup>
 </Project>
