+++ src/System.Reflection/tests/CoreCLR/AssemblyTests.cs
