--- src/System.Runtime.Serialization.Xml/tests/DataContractSerializer.CoreCLR.cs
@@ -1,20 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using SerializationTypes.CoreCLR;
-using System;
-using System.IO;
-using System.Runtime.Serialization;
-using Xunit;
-
-public static partial class DataContractSerializerTests
-{
-    [Fact]
-    public static void DCS_DifferentCollectionsOfSameTypeAsKnownTypes()
-    {
-        Assert.Throws<InvalidOperationException>(() => {
-            (new DataContractSerializer(typeof(TypeWithKnownTypesOfCollectionsWithConflictingXmlName))).WriteObject(new MemoryStream(), new TypeWithKnownTypesOfCollectionsWithConflictingXmlName());
-        });
-    }
-}
