+++ src/System.Runtime.WindowsRuntime/src/System.Runtime.WindowsRuntime.csproj
@@ -33,12 +33,12 @@
     <ProjectReference Include="..\..\System.ObjectModel\src\System.ObjectModel.csproj" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetsNETCoreApp)'=='true' or '$(TargetGroup)'=='uap'">
-    <Compile Include="System\InternalHelpers.CoreCLR.cs" />
-    <Compile Include="System\IO\StreamOperationAsyncResult.CoreCLR.cs" />
+    <Compile Include="System\InternalHelpers.cs" />
+    <Compile Include="System\IO\StreamOperationAsyncResult.cs" />
     <Compile Include="System\Runtime\InteropServices\WindowsRuntime\MarshalingHelpers.cs" />
-    <Compile Include="System\Threading\Tasks\AsyncInfoToTaskBridge.CoreCLR.cs" />
+    <Compile Include="System\Threading\Tasks\AsyncInfoToTaskBridge.cs" />
     <Compile Include="System\Threading\WindowsRuntimeSynchronizationContext.cs" />
-    <Compile Include="System\WindowsRuntimeSystemExtensions.CoreCLR.cs" />
+    <Compile Include="System\WindowsRuntimeSystemExtensions.cs" />
     <Compile Include="$(CommonPath)\Interop\Windows\Mincore\Interop.RoGetActivationFactory.cs">
       <Link>Common\Interop\Windows\Mincore\Interop.RoGetActivationFactory.cs</Link>
     </Compile>
