+++ src/System.Runtime.WindowsRuntime/src/System/InternalHelpers.cs
