+++ src/System.Runtime.WindowsRuntime/src/System/WindowsRuntimeSystemExtensions.cs
